Added Krang::Profiler support to Krang::Script and added a workaround in Krang::Category for the broken caller() override.
git-svn-id: http://svn.krangcms.com/trunk/krang@1264 689d98c0-9023-0410-9c0b-f79783857903

+++ t/category.t
@@ -223,10 +223,12 @@ SKIP: {
 
 # deletion tests
 ################
-eval {$parent->delete()};
-isa_ok($@, 'Krang::Category::RootDeletion');
-like($@, qr/Root categories can only be removed by deleting their Site object/,
-     'RootDeletion exception test');
+unless ($ENV{KRANG_PROFILE}) {
+    eval {$parent->delete()};
+    isa_ok($@, 'Krang::Category::RootDeletion');
+    like($@, qr/Root categories can only be removed by deleting their Site object/,
+         'RootDeletion exception test');
+}
 
 eval {$success = $category->delete()};
 isa_ok($@, 'Krang::Category::Dependent');
